# specs/001-software-architecture-diagrams/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ADR Diagram API - Architecture Diagrams
  version: 1.0.0
paths:
  /diagrams:
    get:
      summary: List active saved-diagram summaries
      description: Returns the active diagrams for the saved-diagram view. Each summary includes a stable ID, name, lifecycle state, and last-saved time; names are not unique selection keys.
      responses:
        '200': {description: Active diagrams, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramSummaryList'}}}}
    post:
      summary: Create a diagram
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramCreate'}}}}
      responses:
        '201': {description: Created diagram, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramDocument'}}}}
        '422': {$ref: '#/components/responses/ValidationError'}
  /diagrams/trash:
    get:
      summary: List trashed diagrams
      responses: {'200': {description: Trashed diagrams, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramSummaryList'}}}}}
  /diagrams/{diagramId}:
    get:
      summary: Load a selected saved diagram document
      description: Returns the complete current saved document for the selected stable diagram ID. Clients must retain the current active document when this request fails.
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Diagram document, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramDocument'}}}}, '404': {description: Not found}}
    put:
      summary: Validate and autosave the complete diagram document
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramWrite'}}}}
      responses: {'200': {description: Saved document, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramDocument'}}}}, '422': {$ref: '#/components/responses/ValidationError'}}
    delete:
      summary: Move a diagram to trash
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'204': {description: Moved to trash}, '404': {description: Not found}}
  /diagrams/{diagramId}/restore:
    post:
      summary: Restore a trashed diagram
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Restored document, content: {application/json: {schema: {$ref: '#/components/schemas/DiagramDocument'}}}}, '409': {description: Diagram is not trashed}}
  /diagrams/{diagramId}/components/{componentId}/dependencies:
    get:
      summary: Count relationships that would be removed with a component
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}, {name: componentId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Dependent relationship count, content: {application/json: {schema: {type: object, required: [relationshipCount], properties: {relationshipCount: {type: integer, minimum: 0}}}}}}, '404': {description: Diagram or component not found}}
  /diagrams/{diagramId}/components/{componentId}:
    delete:
      summary: Remove an unconnected component after client confirmation
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}, {name: componentId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Updated document after removing an unconnected component, content: {application/json: {schema: {type: object, required: [document, relationshipCount], properties: {document: {$ref: '#/components/schemas/DiagramDocument'}, relationshipCount: {type: integer, minimum: 0, maximum: 0}}}}}}, '404': {description: Diagram or component not found}, '409': {$ref: '#/components/responses/ComponentDependencyConflict'}}
  /diagrams/{diagramId}/export/mermaid:
    get:
      summary: Export the validated diagram as Mermaid
      parameters: [{name: diagramId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200':
          description: Mermaid source download
          headers:
            Content-Disposition: {schema: {type: string}, description: Attachment filename ending in .mmd}
          content: {text/vnd.mermaid: {schema: {type: string}}}
        '422': {$ref: '#/components/responses/ValidationError'}
components:
  responses:
    ValidationError:
      description: Actionable validation failure
      content:
        application/json:
          schema:
            type: object
            required: [message, fields]
            properties:
              message: {type: string}
              fields: {type: object, additionalProperties: {type: string}}
    ComponentDependencyConflict:
      description: Component has attached relationships and cannot be removed
      content:
        application/json:
          schema:
            type: object
            required: [message, componentId, relationshipCount]
            properties:
              message: {type: string}
              componentId: {type: string, format: uuid}
              relationshipCount: {type: integer, minimum: 1}
  schemas:
    DiagramCreate:
      type: object
      required: [name]
      properties: {name: {type: string, minLength: 1}}
    DiagramWrite:
      type: object
      required: [name, components, relationships]
      properties:
        name: {type: string, minLength: 1}
        components: {type: array, items: {$ref: '#/components/schemas/Component'}, uniqueItems: true}
        relationships: {type: array, items: {$ref: '#/components/schemas/Relationship'}, uniqueItems: true}
    DiagramDocument:
      allOf:
        - {$ref: '#/components/schemas/DiagramWrite'}
        - type: object
          required: [id, status, createdAt, updatedAt]
          properties:
            id: {type: string, format: uuid}
            status: {type: string, enum: [active, trashed]}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
    Component:
      type: object
      required: [id, name, position]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, minLength: 1}
        description: {type: string, nullable: true}
        type: {type: string, nullable: true}
        position: {type: object, required: [x, y], properties: {x: {type: number}, y: {type: number}}}
    Relationship:
      type: object
      required: [id, sourceComponentId, targetComponentId, direction]
      properties:
        id: {type: string, format: uuid}
        sourceComponentId: {type: string, format: uuid}
        targetComponentId: {type: string, format: uuid}
        direction: {type: string, enum: [directed, undirected]}
        label: {type: string, nullable: true}
    DiagramSummaryList:
      type: array
      items:
        type: object
        required: [id, name, status, updatedAt]
        properties:
          id: {type: string, format: uuid}
          name: {type: string}
          status: {type: string, enum: [active, trashed]}
          updatedAt: {type: string, format: date-time}